refactor: grouping now-playing components
