+++ src/components/now-playing/index.ts
@@ -1,0 +1,9 @@
+import DevicePopover from './device-popover'
+import NowPlaying from './now-playing'
+import NowPlayingBar from './now-playing-bar'
+import NowPlayingCard from './now-playing-card'
+import PlayerPopover from './player-popover'
+
+export { DevicePopover, NowPlayingBar, NowPlayingCard, PlayerPopover }
+
+export default NowPlaying
